requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2019.9.11
chardet==3.0.4
click==8.1.7
colorzero==1.1
Flask==3.0.3
Flask-SocketIO==5.3.6
gpiozero==1.5.1
h11==0.14.0
idna==2.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pigpio==1.44
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.22.0
simple-websocket==1.0.0
urllib3==1.25.3
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
